Move all API methods into  a single file
